requirements.txt
streamlit>=1.32.0
opencv-python>=4.9.0
openai-whisper>=20231117
deepface>=0.0.89
transformers>=4.38.0
spacy>=3.7.0
librosa>=0.10.1
plotly>=5.19.0
reportlab>=4.1.0
sqlalchemy>=2.0.0
mediapipe>=0.10.9
vaderSentiment>=3.3.1
wordcloud>=1.9.3
pyaudio>=0.2.14
torch>=2.2.0
numpy>=1.26.0
pandas>=2.2.0
python-dotenv>=1.0.1
Pillow>=10.2.0
scikit-learn>=1.4.0
SpeechRecognition>=3.10.0
audio-recorder-streamlit>=0.0.8
